requirements.txt
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
extra-streamlit-components==0.1.60
fastapi==0.104.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.1
orjson==3.9.10
packaging==23.2
pandas==2.1.2
passlib==1.7.4
Pillow==10.1.0
protobuf==4.24.4
psycopg2-binary==2.9.9
py-postgresql==1.3.0
pyarrow==13.0.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
streamlit==1.28.0
streamlit-authenticator==0.2.3
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.19.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
websockets==12.0
zipp==3.17.0
